requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
#torch==1.10.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
#torchvision>=0.11.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
tqdm>=4.41.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
onnx>=1.9.0  # ONNX export
onnx-simplifier>=0.3.6  # ONNX simplifier
onnxruntime>=1.9.0  # ONNX runtime
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export
mlsocket>=0.1.0  # ML Socket export for transmitting
# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
albumentations>=1.0.3
# Torch must be 1.10.0+cu113 or above. This is the version that I've tested with.
torch==1.13.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
# pycocotools>=2.0  # COCO mAP
#roboflow
